requirements.txt

opencv-python>=4.5.0
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
Pillow>=8.0.0
ultralytics>=8.0.0


scikit-image>=0.19.0
imageio>=2.9.0
opencv-contrib-python>=4.5.0
albumentations>=1.0.0


pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0


scikit-learn>=0.24.0
tqdm>=4.62.0
PyYAML>=5.4.0


tensorboard>=2.6.0
pytorch-lightning>=1.5.0


python-dotenv>=0.19.0
requests>=2.26.0
